Fixed a bug in the KD-Tree building algorithm causing infinite recursion when a node contains all the candidate polygons.

+++ src/BoundingBox.h
@@ -44,8 +44,8 @@ struct BoundingBox
         return max.z() - min.z();
     }
 
-    float volume () const
+    double volume () const
     {
-        return width() * height() * depth();
+        return (double)width() * height() * depth();
     }
 };
